+++ src/azurestorage.hpp
@@ -21,8 +21,7 @@
 #include <string>
 #include <vector>
 
-#include <spdlog/spdlog.h>
-
+#include "logging.hpp"
 #include "status.hpp"
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Wall"
