+++ src/filesystem.hpp
@@ -22,8 +22,7 @@
 #include <string>
 #include <vector>
 
-#include <spdlog/spdlog.h>
-
+#include "logging.hpp"
 #include "model_version_policy.hpp"
 #include "openssl/md5.h"
 #include "status.hpp"
@@ -133,7 +132,7 @@ public:
         std::string file_template = "/tmp/fileXXXXXX";
         char* tmp_folder = mkdtemp(const_cast<char*>(file_template.c_str()));
         if (tmp_folder == nullptr) {
-            SPDLOG_ERROR("Failed to create local temp folder: {} {}", file_template, strerror(errno));
+            SPDLOG_LOGGER_ERROR(modelmanager_logger, "Failed to create local temp folder: {} {}", file_template, strerror(errno));
             return StatusCode::FILESYSTEM_ERROR;
         }
         fs::permissions(tmp_folder,
@@ -262,7 +261,7 @@ public:
         int status =
             mkdir(const_cast<char*>(path.c_str()), S_IRUSR | S_IWUSR | S_IXUSR);
         if (status == -1) {
-            SPDLOG_ERROR("Failed to create local folder: {} {} ", path, strerror(errno));
+            SPDLOG_LOGGER_ERROR(modelmanager_logger, "Failed to create local folder: {} {} ", path, strerror(errno));
             return StatusCode::PATH_INVALID;
         }
         return StatusCode::OK;
