+++ src/python/ovms_py_tensor_py.cpp
@@ -14,10 +14,10 @@
 // limitations under the License.
 //*****************************************************************************
 
-#include "../../python/ovms_py_tensor.hpp"
-
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
+
+#include "ovms_py_tensor.hpp"
 
 namespace py = pybind11;
 using namespace ovms;
