+++ src/python/pythoninterpretermodule.hpp
@@ -1,0 +1,45 @@
+//****************************************************************************
+// Copyright 2023 Intel Corporation
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//*****************************************************************************
+#pragma once
+#include <memory>
+#include <thread>
+
+#include "../module.hpp"
+
+namespace pybind11 {
+class gil_scoped_release;
+}
+namespace py = pybind11;
+
+namespace ovms {
+class Config;
+class PythonBackend;
+
+class PythonInterpreterModule : public Module {
+    PythonBackend* pythonBackend{nullptr};
+    mutable std::unique_ptr<py::gil_scoped_release> GILScopedRelease;
+    std::thread::id threadId;
+
+public:
+    PythonInterpreterModule();
+    ~PythonInterpreterModule();
+    Status start(const ovms::Config& config) override;
+    void shutdown() override;
+    PythonBackend* getPythonBackend() const;
+    void releaseGILFromThisThread() const;
+    void reacquireGILForThisThread() const;
+};
+}  // namespace ovms
