+++ src/python/pythonnoderesources.hpp
@@ -21,7 +21,10 @@
 
 #include <pybind11/embed.h>  // everything needed for embedding
 
-#include "src/mediapipe_calculators/python_executor_calculator_options.pb.h"
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wdeprecated-declarations"
+#include "mediapipe/framework/calculator_graph.h"
+#pragma GCC diagnostic pop
 
 namespace py = pybind11;
 
@@ -47,5 +50,5 @@ public:
 private:
     static py::dict preparePythonNodeInitializeArguments(const ::mediapipe::CalculatorGraphConfig::Node& graphNodeConfig);
 };
-
+using PythonNodeResourcesMap = std::unordered_map<std::string, std::shared_ptr<PythonNodeResources>>;
 }  // namespace ovms
