+++ src/server.hpp
@@ -21,6 +21,7 @@
 #include <unordered_map>
 
 #include "module.hpp"
+#include "module_names.hpp"
 
 namespace ovms {
 class Config;
@@ -28,13 +29,6 @@ class Status;
 
 class ServerSettingsImpl;
 class ModelsSettingsImpl;
-
-extern const std::string PROFILER_MODULE_NAME;
-extern const std::string GRPC_SERVER_MODULE_NAME;
-extern const std::string HTTP_SERVER_MODULE_NAME;
-extern const std::string SERVABLE_MANAGER_MODULE_NAME;
-extern const std::string METRICS_MODULE_NAME;
-extern const std::string PYTHON_INTERPRETER_MODULE_NAME;
 
 class Server {
     mutable std::shared_mutex modulesMtx;
